Live profile plot day + energy coop rename + and rb pointer fixes (#38)
* Added day live plot option

A new radio-button appears when viewing the live profile plots to switch the consumption/production plots between 1 week and 1 day x-axis scale

* Function to enable/disable non-live radiobuttons

f_setNonLivePlotRadioButtons takes a boolean argument. When true is passed, it enables the radiobuttons to switch away from the live plot. This is done after a rapid run. When false is passed it disables them. This is done when f_resetSettings is called in the interface

* energycoop rename + fix of variable radio button pointers.

---------

Co-authored-by: Luc-Sol <luc@solidegroep.nl>